# docs/openapi.yaml
openapi: 3.0.3
info:
  title: automotive-workshop-api — Customer & Vehicle Management
  description: >
    Auth, Customer Management, Vehicle Management, Service catalog, Product &
    Stock Management, Service Order opening/diagnosis/quote/execution, Service
    Order listing/detail (specs/service-order-query/), and Service Order
    Tracking (specs/service-order-tracking/) endpoints for the automotive
    workshop API. See specs/auth/, specs/customer-management/,
    specs/vehicle-management/, specs/service-catalog/,
    specs/product-management/, specs/service-order-opening/,
    specs/service-order-diagnosis-quote/, specs/service-order-quote-decision/,
    specs/service-order-execution/, specs/service-order-query/, and
    specs/service-order-tracking/ for the requirements and design behind
    these contracts. Auth, Vehicle, Service catalog, Product, and Service
    Order Query/Diagnosis/Quote/Execution routes require a JWT (see the
    bearerAuth security scheme; POST /api/v1/auth/login itself is the one
    exception, since it is what issues the token); Customer routes and
    Service Order creation do not (see
    specs/customer-management/requirements.md §7.2 and CLAUDE.md §1); Service
    Order Tracking and the customer-facing quote approve/reject endpoints use
    neither — they validate their own per-order tracking token instead (see
    the trackingToken security scheme).
  version: "1.0.0"
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/produtos:
    post:
      tags: [products]
      summary: Register a new product (part or supply)
      operationId: createProduct
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
            examples:
              part:
                summary: A replacement part
                value:
                  name: Filtro de Óleo
                  description: Filtro de óleo compatível com motores 1.0/1.6.
                  unitPrice: 35.9
                  currentStock: 50
                  type: PART
      responses:
        "201":
          description: The created product, always ACTIVE.
          headers:
            Location:
              description: URL of the newly created product.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: >
            Malformed JSON body (code `INVALID_BODY`), or a missing/invalid
            field — empty name, missing/negative unitPrice, missing/negative
            currentStock, non-positive code, or type not PART/SUPPLY (code
            `VALIDATION_ERROR`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing, malformed, or expired bearer token.
        "409":
          description: The supplied code already belongs to another product (code `DUPLICATE_CODE`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    get:
      tags: [products]
      summary: List products (paginated, filterable)
      operationId: listProducts
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: type
          in: query
          description: Filter by product type.
          schema:
            type: string
            enum: [PART, SUPPLY]
        - name: status
          in: query
          description: Filter by product status.
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
      responses:
        "200":
          description: A page of products.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
        "401":
          description: Missing, malformed, or expired bearer token.
  /api/v1/produtos/{id}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      tags: [products]
      summary: Retrieve a product by id
      operationId: getProductById
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: "{id} is not a valid UUID (code `INVALID_BODY`)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: No product with this id (code `NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      tags: [products]
      summary: Partially update a product's cadastral data
      description: >
        Name, description, unitPrice, and type may be updated. `currentStock`
        is deliberately not modifiable through this endpoint (RNF07) — a
        request that includes it is rejected wholesale, even alongside other,
        otherwise-valid fields; use POST .../estoque/ajustes instead.
      operationId: updateProduct
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: The updated product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: >
            Malformed JSON body, `currentStock` present (code
            `VALIDATION_ERROR`, RNF07), empty name, negative unitPrice, or
            type not PART/SUPPLY.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: No product with this id (code `NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [products]
      summary: Logically deactivate a product
      description: >
        Soft delete: sets status to INACTIVE. Idempotent. An inactive
        product must not be used in new quotes or stock adjustments.
      operationId: deactivateProduct
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The now-inactive product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: No product with this id (code `NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/produtos/{id}/estoque/ajustes:
    post:
      tags: [products]
      summary: Register a stock adjustment (entry or exit)
      description: >
        Applied atomically at the database level (RNF07): an EXIT is
        rejected as 409 if it would take `currentStock` negative, closing
        the race between two concurrent adjustments (RN01). The product must
        be ACTIVE.
      operationId: adjustProductStock
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustmentRequest"
            examples:
              entry:
                summary: Stock entry
                value:
                  type: ENTRY
                  quantity: 10
                  reason: Ajuste de inventário anual
      responses:
        "201":
          description: The created stock movement record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductStockMovement"
        "400":
          description: >
            Malformed JSON body, type not ENTRY/EXIT, non-positive quantity,
            or empty reason (code `VALIDATION_ERROR`, RN01).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: No product with this id (code `NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: >
            An EXIT would take `currentStock` negative (code
            `INSUFFICIENT_STOCK`, RN01), or the product is inactive (code
            `VALIDATION_ERROR`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/produtos/{id}/estoque:
    get:
      tags: [products]
      summary: Retrieve a product's current stock balance
      operationId: getProductStockBalance
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: The product's stock balance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductStockBalance"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: No product with this id (code `NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/produtos/{id}/movimentacoes:
    get:
      tags: [products]
      summary: List a product's stock movement history (RF10)
      operationId: listProductStockMovements
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ProductId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of the product's stock movements, most recent first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductStockMovementList"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: No product with this id (code `NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /api/v1/acompanhamento/{codigo}:
    get:
      tags: [tracking]
      summary: Look up a service order's public tracking view
      description: >
        Customer-facing read view (RF12) — never accepts or requires the
        administrative JWT (specs/service-order-tracking/requirements.md
        AC10); authenticates with the X-Tracking-Token header instead (see
        the trackingToken security scheme). The response is a deliberately
        limited projection, distinct from the administrative
        ServiceOrderDetail: only the order's code, current status, a partial
        vehicle identification (license plate, brand, model, year, color —
        no id/code/status/owning customer), and the milestone trail (event,
        previous/new status, occurredAt — no free-text description). Customer
        PII (CPF/CNPJ, phone, e-mail), internal ids, and any other order are
        never present (AC6-AC9).
      operationId: getServiceOrderTracking
      security:
        - trackingToken: []
      parameters:
        - name: codigo
          in: path
          required: true
          description: The service order's sequential code.
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: The service order's public tracking view.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrderTracking"
        "401":
          description: >
            The X-Tracking-Token header is missing, malformed, wrong,
            belongs to a different service order, or has been revoked (code
            `INVALID_TRACKING_TOKEN`) — all deliberately indistinguishable
            from each other, so a wrong token never reveals which specific
            way it was wrong (AC2, AC4, AC5).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: >
            No service order with this code, or {codigo} is not a valid
            integer (code `NOT_FOUND`) — checked before the token, so it
            takes precedence regardless of what token was supplied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Required on every Vehicle Management route, every Service catalog and
        Product Management route, GET /api/v1/auth/me, and every Service
        Order route except creation (RNF02) — that includes the Service
        Order Query routes plus the diagnosis/quote/execution routes added by
        specs/service-order-diagnosis-quote/,
        specs/service-order-quote-decision/ (only its send-quote endpoint;
        approve/reject are customer-facing, see trackingToken), and
        specs/service-order-execution/. Not required on Customer Management
        routes (see specs/customer-management/requirements.md §7.2) nor on
        POST /api/v1/service-orders (CLAUDE.md §1), and not accepted at all
        by the Service Order Tracking route or the customer-facing quote
        approve/reject routes (see trackingToken).
    trackingToken:
      type: apiKey
      in: header
      name: X-Tracking-Token
      description: >
        Per-service-order token, issued once as the trackingToken field of
        POST /api/v1/service-orders's creation response (there is no
        separate issuance endpoint). Deliberately a distinct header, not
        Authorization: Bearer, so it can never be confused with the
        administrative JWT (specs/service-order-tracking/requirements.md §0).
  parameters:
...
    ProductId:
      name: id
      in: path
      required: true
      description: The product's technical UUID.
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      description: 1-indexed page number. Defaults to 1; values below 1 fall back to the default.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      description: Items per page. Defaults to 20; clamped to a maximum of 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    ServiceOrderTrackingVehicle:
      type: object
      description: >
        Deliberately partial — no id, code, status, or owning customer
        (specs/service-order-tracking/requirements.md AC6).
      properties:
        licensePlate:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
    ServiceOrderTrackingMilestone:
      type: object
      description: >
        A status-change event, without the internal id or free-text
        description an administrative history entry carries (AC6).
      properties:
        event:
          type: string
          enum:
            [
              creation,
              diagnosis_started,
              quote_composed,
              quote_sent,
              approval,
              completion,
              cancellation,
              delivery,
            ]
        previousStatus:
          type: string
        newStatus:
          type: string
        occurredAt:
          type: string
          format: date-time
    ServiceOrderTracking:
      type: object
      description: >
        The public tracking contract — a distinct type from the
        administrative ServiceOrderDetail, with no shared fields beyond code
        and status (AC9). Never contains CPF/CNPJ, phone, e-mail, internal
        ids, or any other service order (AC7, AC8).
      properties:
        code:
          type: integer
          format: int64
        status:
          type: string
          enum:
            [
              RECEIVED,
              IN_DIAGNOSIS,
              AWAITING_APPROVAL,
              IN_PROGRESS,
              COMPLETED,
              DELIVERED,
              CANCELED,
            ]
        vehicle:
          $ref: "#/components/schemas/ServiceOrderTrackingVehicle"
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/ServiceOrderTrackingMilestone"
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                [
                  INVALID_BODY,
                  VALIDATION_ERROR,
                  NOT_FOUND,
                  DUPLICATE_DOCUMENT,
                  DUPLICATE_EMAIL,
                  CUSTOMER_NOT_FOUND,
                  CUSTOMER_INACTIVE,
                  DUPLICATE_LICENSE_PLATE,
                  INVALID_TRACKING_TOKEN,
                  INVALID_REQUEST,
                  UNAUTHORIZED,
                  CODE_ALREADY_EXISTS,
                  DUPLICATE_CODE,
                  INSUFFICIENT_STOCK,
                  VEHICLE_INACTIVE,
                  VEHICLE_NOT_OWNED_BY_CUSTOMER,
                  INVALID_STATUS_TRANSITION,
                  DIAGNOSIS_NOT_STARTED,
                  QUOTE_ALREADY_DECIDED,
                  PRODUCT_INACTIVE,
                  SERVICE_EXECUTION_ALREADY_FINISHED,
                  EXECUTIONS_NOT_COMPLETED,
                  INTERNAL_ERROR,
                ]
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
      example:
        error:
          code: VALIDATION_ERROR
          message: invalid customer data
          details:
            - field: document
              message: "CPF \"12345678900\" has invalid check digits"
...
    CreateProductRequest:
      type: object
      required: [name, unitPrice, currentStock, type]
      properties:
        code:
          type: integer
          format: int64
          nullable: true
          description: Optional. When omitted, the database generates the sequential code.
        name:
          type: string
        description:
          type: string
        unitPrice:
          type: number
          format: double
          description: Required. Must not be negative.
        currentStock:
          type: integer
          description: Required. Must not be negative.
        type:
          type: string
          enum: [PART, SUPPLY]
    UpdateProductRequest:
      type: object
      description: >
        Every field is optional; omitted fields keep their current value.
        `currentStock` is never part of this request's shape — it is
        immutable through this endpoint (RNF07); use
        POST .../estoque/ajustes instead.
      properties:
        name:
          type: string
        description:
          type: string
        unitPrice:
          type: number
          format: double
        type:
          type: string
          enum: [PART, SUPPLY]
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        unitPrice:
          type: number
          format: double
        currentStock:
          type: integer
        type:
          type: string
          enum: [PART, SUPPLY]
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    StockAdjustmentRequest:
      type: object
      required: [type, quantity, reason]
      properties:
        type:
          type: string
          enum: [ENTRY, EXIT]
        quantity:
          type: integer
          description: Must be greater than zero.
        reason:
          type: string
    ProductStockBalance:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: integer
          format: int64
        name:
          type: string
        currentStock:
          type: integer
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        updatedAt:
          type: string
          format: date-time
    ProductStockMovement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        type:
          type: string
          enum: [ENTRY, EXIT]
        quantity:
          type: integer
        previousStock:
          type: integer
        newStock:
          type: integer
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
    ProductStockMovementList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProductStockMovement"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...